requirements.txt
tweepy>=3.8.0
google-cloud-language>=1.3.0
redis>=3.3.8
Flask>=1.1.1
Flask-Cors>=3.0.8
uWSGI>=2.0.17.1